Covid19 Data Analysis Using Python
Now we will build (project goals). We will accomplish it in by completing each task in the project:

Introduction     

Importing the auto-update dataset using Github url

Creating a choropleth map to visualize the global spread of the virus

Creating a bar chart to compare countries in terms of the intensity of the COVID19 transmission   

Visualizing the spread of the virus Before and After lockdown in Italy

While you are watching me work on each step, you will get a cloud desktop with all the required software pre-installed. This will allow you to follow along the instructions to complete the above mentioned tasks. After all, we learn best with active, hands-on learning

Ready to get started? Click on the button below to launch the project on Rhyme.